Simplify a bunch of stuff...
