+++ package.json
@@ -13,7 +13,6 @@
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
     "react-is": "^17.0.2",
-    "react-pdf": "^5.3.2",
     "react-scripts": "4.0.3",
     "styled-components": "^5.3.1"
   },
@@ -27,6 +26,7 @@
     "extends": [
       "react-app",
       "react-app/jest",
+      "eslint:recommended",
       "plugin:import/recommended",
       "plugin:import/typescript"
     ],
@@ -51,7 +51,6 @@
     "@types/node": "^16.7.8",
     "@types/react": "^17.0.0",
     "@types/react-dom": "^17.0.0",
-    "@types/react-pdf": "^5.0.5",
     "@types/styled-components": "^5.1.11",
     "customize-cra": "^1.0.0",
     "eslint": "^7.32.0",
